blogilistan laajennus, step 10 done

+++ controllers/blogs.js
@@ -33,8 +33,18 @@ blogsRouter.get('/:id', async (request, response, next) => {
 // Delete a blog post
 blogsRouter.delete('/:id', async (request, response, next) => {
   try {
-    await Blog.findByIdAndRemove(request.params.id)
-    response.status(204).end()
+    const decodedToken = jwt.verify(request.token, process.env.SECRET)
+    if (!request.token || !decodedToken.id) {
+      return response.status(401).json({ error: 'token missing or invalid' })
+    }
+
+    const user = await User.findById(decodedToken.id)
+
+    const delBlog = await Blog.findById(request.params.id)
+    if ( delBlog.user.toString() === user.id.toString() ) {
+      await Blog.findByIdAndRemove(request.params.id)
+      response.status(204).end()
+    }
   } catch (exception) {
     next(exception)
   }
